+++ apps/www/src/models/TagsFilterType.ts
@@ -1,0 +1,2 @@
+import { Tag } from "@phylopic/utils"
+export type TagsFilterType = ReadonlySet<Tag> | undefined