+++ packages/api-models/src/queryParameters/types/NodeLineageParameters.ts
@@ -1,3 +1,6 @@
+import { UUID } from "@phylopic/utils"
 import { NodeEmbedded } from "../../types/NodeWithEmbedded"
 import { ListParameters } from "./ListParameters"
-export type NodeLineageParameters = ListParameters<NodeEmbedded>
+export type NodeLineageParameters = ListParameters<NodeEmbedded> & {
+    uuid: UUID
+}