+++ packages/utils/src/models/normalization/index.ts
@@ -1,3 +1,4 @@
 export * from "./normalizeNomen"
 export * from "./normalizeNomina"
+export * from "./normalizeTag"
 export * from "./normalizeUUID"